move typescript extraction rules behind the indexer language registry
S3-TYPESCRIPT (ARCH-012), following S3-KOTLIN (#1676). TypeScript's tree-sitter
graph and tags rule files, compiled-rule cache, member-call callsite marker,
and receiver-call resolution engine move out of the indexer god file into
`codestory-indexer/src/languages/typescript.rs` as one `LanguageExtraction`
row. Every language-keyed dispatch that spelled "typescript" now reads the
registry first and keeps its residual `match` for the languages that have not
moved; the last sibling package deletes those arms.

The move is mechanical: 20 functions left `lib.rs` and the same 20 landed in
`languages/typescript.rs` byte-identically, apart from renaming the two entry
points (`collect_typescript_receiver_call_edges` -> `receiver_call_specs`,
`typescript_member_call` -> `member_call`) and widening two of them to
`pub(crate)`. Rebuilding the module as `header + renamed(deleted lines)` and
running rustfmt reproduces the committed file with an empty diff. The module
adds one function, the grammar constructor `typescript_language`, which
replaces `language_configs::typescript`.

`collect_typescript_imported_type_bindings` and
`typescript_property_belongs_to_owner` are `pub(crate)` because JavaScript's
still-unmigrated receiver-call engine calls them. They are TypeScript rules
that JavaScript borrows rather than shared crate-root helpers, so they move
with the rest and the two JavaScript call sites in `lib.rs` reach them by path
until #1680 lands. JavaScript's own logic is untouched.

TSX stays a separate rollback unit (#1682). It keeps its own grammar, graph
rule file, compiled-rule cache, dispatch name and `language_configs` arm; this
row deliberately claims only `ts`/`mts`/`cts` and the dispatch name
"typescript". The two places where TSX shared a TypeScript asset now reach it
through this module: `TSX_TAGS_QUERY` and the `"tsx"` receiver-call arm.

Proof: `tests/language_extraction_snapshot.rs` gains one row plus two goldens
pinning the full rendered projection of both TypeScript fixtures. The goldens
were captured before the production edit and are byte-identical after
(sha256 0c76d00f.. / 03268e20..). Falsified by mutation: `receiver_call_specs:
None` and `qualified_name_delimiter: "::"` each turn both snapshots red, and
the delimiter mutation leaves `fidelity_regression` and
`tictactoe_language_coverage` fully green — the threshold suites alone would
wave that regression through.

Declared behavior changes:
- None to any indexing, resolution, or rendering output. TypeScript
  fidelity_regression and tictactoe_language_coverage projections are proven
  byte-identical before and after, and packet_search_eval is unmoved.
- `LanguageRuleset::TypeScript`'s arm in `compiled_rules` is now unreachable
  and returns a typed error instead of compiled rules; the registry answers
  first. Fail-closed by construction, matching the Kotlin arm.
- `member_callsite_marker_for_call_syntax("ts_member")` now answers for `.tsx`
  files too, because `tsx.graph.scm` emits the same `call_syntax` value as
  `typescript.graph.scm`. It returns the same string the deleted
  `TS_MEMBER_CALLSITE_MARKER` arm returned, so the emitted callsite identity is
  unchanged.
- New public row in `codestory_contracts::language_support`:
  `LANGUAGE_COMMENT_PROFILES` gains `typescript`. `line_comment_for_language`
  answers `Some("//")` for it, exactly what the CLI roster returned; `tsx` has
  no public language profile and keeps answering from the CLI's own roster.
- `LANGUAGE_SUPPORT_PROFILES` is unchanged, so extension routing, claim tiers,
  `supported_extensions`, `is_structural_source_path`, and every existing
  language claim are identical.
- The CLI comment roster and the indexer route-comment roster still disagree by
  design (CLI has cpp/swift/tsx/jsx and not c; the route scanner has vue/astro
  and not cpp/swift/tsx). Only TypeScript's own entry moved in each.
- No version surface touched; everything stays 0.16.3.

Closes #1681
