+++ crates/codestory-indexer/src/language_configs.rs
@@ -2,8 +2,7 @@ use super::{
     BASH_GRAPH_QUERY, C_GRAPH_QUERY, CPP_GRAPH_QUERY, CSHARP_GRAPH_QUERY, DART_GRAPH_QUERY,
     GO_GRAPH_QUERY, JAVA_GRAPH_QUERY, JAVASCRIPT_GRAPH_QUERY, LanguageConfig, LanguageRuleset,
     PHP_GRAPH_QUERY, PYTHON_GRAPH_QUERY, RUBY_GRAPH_QUERY, RUST_GRAPH_QUERY, RUST_TAGS_QUERY,
-    SWIFT_GRAPH_QUERY, TSX_GRAPH_QUERY, TSX_TAGS_QUERY, TYPESCRIPT_GRAPH_QUERY,
-    TYPESCRIPT_TAGS_QUERY, languages, make_language_config,
+    SWIFT_GRAPH_QUERY, TSX_GRAPH_QUERY, TSX_TAGS_QUERY, languages, make_language_config,
 };
 use codestory_contracts::language_support::{
     LanguageSupportMode, language_support_profile_for_ext, normalize_extension,
@@ -33,8 +32,9 @@ pub(super) fn get_language_for_ext(ext: &str) -> Option<LanguageConfig> {
         ("java", _) => Some(java()),
         ("rust", _) => Some(rust()),
         ("javascript", _) => Some(javascript()),
+        // `ts`/`mts`/`cts` are answered by the registry above; `tsx` keeps its
+        // own grammar and rule file until #1682 gives it a registry row.
         ("typescript", "tsx") => Some(tsx()),
-        ("typescript", _) => Some(typescript()),
         ("cpp", _) => Some(cpp()),
         ("c", _) => Some(c()),
         ("go", _) => Some(go()),
@@ -85,16 +85,6 @@ fn javascript() -> LanguageConfig {
         JAVASCRIPT_GRAPH_QUERY,
         None,
         LanguageRuleset::JavaScript,
-    )
-}
-
-fn typescript() -> LanguageConfig {
-    make_language_config(
-        tree_sitter_typescript::LANGUAGE_TYPESCRIPT.into(),
-        "typescript",
-        TYPESCRIPT_GRAPH_QUERY,
-        Some(TYPESCRIPT_TAGS_QUERY),
-        LanguageRuleset::TypeScript,
     )
 }
