+++ crates/codestory-indexer/src/lib.rs
@@ -65,7 +65,6 @@ pub(crate) const CSHARP_MEMBER_CALLSITE_MARKER: &str = "syntax:csharp-member-cal
 pub(crate) const RUBY_MEMBER_CALLSITE_MARKER: &str = "syntax:ruby-member-call";
 pub(crate) const PHP_MEMBER_CALLSITE_MARKER: &str = "syntax:php-member-call";
 pub(crate) const JS_MEMBER_CALLSITE_MARKER: &str = "syntax:js-member-call";
-pub(crate) const TS_MEMBER_CALLSITE_MARKER: &str = "syntax:ts-member-call";
 pub(crate) const DART_MEMBER_CALLSITE_MARKER: &str = "syntax:dart-member-call";
 pub(crate) const SWIFT_MEMBER_CALLSITE_MARKER: &str = "syntax:swift-member-call";
 pub(crate) const RECEIVER_OWNER_CALLSITE_PREFIX: &str = "receiver-owner:";
@@ -132,10 +131,10 @@ const JAVA_GRAPH_QUERY: &str = include_str!("../rules/java.scm");
 const RUST_GRAPH_QUERY: &str = include_str!("../rules/rust.graph.scm");
 const RUST_TAGS_QUERY: &str = include_str!("../rules/rust.tags.scm");
 const JAVASCRIPT_GRAPH_QUERY: &str = include_str!("../rules/javascript.scm");
-const TYPESCRIPT_GRAPH_QUERY: &str = include_str!("../rules/typescript.graph.scm");
-const TYPESCRIPT_TAGS_QUERY: &str = include_str!("../rules/typescript.tags.scm");
 const TSX_GRAPH_QUERY: &str = include_str!("../rules/tsx.graph.scm");
-const TSX_TAGS_QUERY: &str = TYPESCRIPT_TAGS_QUERY;
+// TSX still reuses TypeScript's tags query verbatim; the asset moved into the
+// TypeScript registry row and #1682 takes ownership of this reference.
+const TSX_TAGS_QUERY: &str = languages::typescript::TAGS_QUERY;
 const CPP_GRAPH_QUERY: &str = include_str!("../rules/cpp.scm");
 const C_GRAPH_QUERY: &str = include_str!("../rules/c.scm");
 const GO_GRAPH_QUERY: &str = include_str!("../rules/go.scm");
@@ -304,12 +303,10 @@ impl LanguageRuleset {
             LanguageRuleset::JavaScript => {
                 compiled_rules_cache(language, JAVASCRIPT_GRAPH_QUERY, None, &JAVASCRIPT_RULES)
             }
-            LanguageRuleset::TypeScript => compiled_rules_cache(
-                language,
-                TYPESCRIPT_GRAPH_QUERY,
-                Some(TYPESCRIPT_TAGS_QUERY),
-                &TYPESCRIPT_RULES,
-            ),
+            // Answered by the registry above; see the Kotlin arm below.
+            LanguageRuleset::TypeScript => Err(anyhow!(
+                "typescript compiled rules are owned by the language registry"
+            )),
             LanguageRuleset::Tsx => {
                 compiled_rules_cache(language, TSX_GRAPH_QUERY, Some(TSX_TAGS_QUERY), &TSX_RULES)
             }
@@ -376,7 +373,6 @@ static PYTHON_RULES: OnceLock<Result<CompiledLanguageRules, String>> = OnceLock:
 static JAVA_RULES: OnceLock<Result<CompiledLanguageRules, String>> = OnceLock::new();
 static RUST_RULES: OnceLock<Result<CompiledLanguageRules, String>> = OnceLock::new();
 static JAVASCRIPT_RULES: OnceLock<Result<CompiledLanguageRules, String>> = OnceLock::new();
-static TYPESCRIPT_RULES: OnceLock<Result<CompiledLanguageRules, String>> = OnceLock::new();
 static TSX_RULES: OnceLock<Result<CompiledLanguageRules, String>> = OnceLock::new();
 static CPP_RULES: OnceLock<Result<CompiledLanguageRules, String>> = OnceLock::new();
 static C_RULES: OnceLock<Result<CompiledLanguageRules, String>> = OnceLock::new();
@@ -6160,7 +6156,7 @@ fn collect_javascript_receiver_call_edges(
 ) -> Vec<ManualReceiverCallSpec> {
     let mut edges = Vec::new();
     let imported_type_bindings =
-        collect_typescript_imported_type_bindings(tree.root_node(), source);
+        languages::typescript::collect_typescript_imported_type_bindings(tree.root_node(), source);
     walk_tree_nodes(tree.root_node(), &mut |callable| {
         if !matches!(
             callable.kind(),
@@ -6487,7 +6483,7 @@ fn javascript_property_receiver_candidate<'tree>(
         ));
     }
     if matches!(node.kind(), "field_definition" | "public_field_definition")
-        && typescript_property_belongs_to_owner(node, class_node)
+        && languages::typescript::typescript_property_belongs_to_owner(node, class_node)
         && !javascript_surface_starts_with_static(node, source)
     {
         let field_name = javascript_class_field_name(node, source)?;
@@ -7594,7 +7590,9 @@ fn language_receiver_call_specs(
     match language_name {
         "javascript" => collect_javascript_receiver_call_edges(tree, source),
         "python" => collect_python_receiver_call_edges(tree, source),
-        "typescript" | "tsx" => collect_typescript_receiver_call_edges(tree, source),
+        // TSX has always run TypeScript's receiver-call engine; the engine
+        // moved into the registry row and #1682 takes over this arm.
+        "tsx" => languages::typescript::receiver_call_specs(tree, source),
         "java" => collect_java_receiver_call_edges(tree, source),
         "go" => collect_go_receiver_call_edges(tree, source),
         "ruby" => collect_ruby_receiver_call_edges(tree, source),
@@ -8848,586 +8846,6 @@ fn collect_python_imported_type_bindings(
         }
     }
     bindings
-}
-
-fn collect_typescript_receiver_call_edges(
-    tree: &Tree,
-    source: &str,
-) -> Vec<ManualReceiverCallSpec> {
-    let mut edges = Vec::new();
-    let imported_type_bindings =
-        collect_typescript_imported_type_bindings(tree.root_node(), source);
-    let namespace_import_bindings =
-        collect_typescript_namespace_import_bindings(tree.root_node(), source);
-    walk_tree_nodes(tree.root_node(), &mut |callable| {
-        if !matches!(
-            callable.kind(),
-            "function_declaration" | "method_definition" | "arrow_function"
-        ) {
-            return;
-        }
-        let Some(source_name) = js_like_callable_source_name(callable, source) else {
-            return;
-        };
-        let call_source = ManualReceiverSource {
-            name: &source_name,
-            span: ts_node_graph_span(callable),
-        };
-        let mut local_receiver_callsites = HashSet::new();
-        collect_typescript_constructor_receiver_call_specs(
-            callable,
-            source,
-            ManualReceiverSource {
-                name: call_source.name,
-                span: call_source.span,
-            },
-            &imported_type_bindings,
-            &namespace_import_bindings,
-            &mut local_receiver_callsites,
-            &mut edges,
-        );
-        let parameter_receiver_types = collect_colon_parameter_types(callable, source);
-        let mut receiver_types = parameter_receiver_types.clone();
-        if let Some(owner_name) = enclosing_node_with_kind(callable, &["class_declaration"])
-            .and_then(|owner| declaration_name(owner, source))
-            && callable.kind() == "method_definition"
-        {
-            receiver_types.insert("this".to_string(), owner_name);
-        }
-        let property_receiver_types = collect_typescript_class_property_receiver_bindings(
-            callable,
-            source,
-            &imported_type_bindings,
-            &namespace_import_bindings,
-        );
-        receiver_types.extend(
-            property_receiver_types
-                .iter()
-                .map(|(receiver_name, (owner_name, _))| {
-                    (receiver_name.clone(), owner_name.clone())
-                }),
-        );
-        if receiver_types.is_empty() {
-            return;
-        }
-        let mut receiver_modules =
-            collect_typescript_parameter_type_modules(callable, source, &namespace_import_bindings);
-        for (receiver_name, (_, owner_module)) in &property_receiver_types {
-            if let Some(module_name) = owner_module {
-                receiver_modules.insert(receiver_name.clone(), module_name.clone());
-            }
-        }
-        let start = edges.len();
-        collect_receiver_call_specs_in_callable(
-            callable,
-            source,
-            ManualReceiverSource {
-                name: call_source.name,
-                span: call_source.span,
-            },
-            &receiver_types,
-            typescript_member_call,
-            false,
-            &mut edges,
-        );
-        let mut fallback_specs = edges.split_off(start);
-        fallback_specs
-            .retain(|spec| !local_receiver_callsites.contains(&receiver_callsite_key(spec)));
-        for spec in &mut fallback_specs {
-            if parameter_receiver_types.contains_key(&spec.receiver_name)
-                && let Some(binding) = imported_type_bindings.get(&spec.owner_name)
-            {
-                spec.owner_name = binding.owner_name.clone();
-                spec.owner_module = Some(binding.module_name.clone());
-            } else if let Some(module_name) = receiver_modules.get(&spec.receiver_name) {
-                spec.owner_module = Some(module_name.clone());
-            }
-        }
-        edges.extend(fallback_specs);
-    });
-    edges
-}
-
-fn collect_typescript_constructor_receiver_call_specs(
-    callable: TsNode<'_>,
-    source: &str,
-    call_source: ManualReceiverSource<'_>,
-    imported_type_bindings: &HashMap<String, ImportedTypeBinding>,
-    namespace_import_bindings: &HashMap<String, String>,
-    local_receiver_callsites: &mut HashSet<ReceiverCallSiteKey>,
-    edges: &mut Vec<ManualReceiverCallSpec>,
-) {
-    walk_tree_nodes(callable, &mut |node| {
-        let Some((receiver_name, method_name)) = typescript_member_call(node, source) else {
-            return;
-        };
-        if !receiver_call_belongs_to_callable(node, callable) {
-            return;
-        }
-        let Some(owner) = typescript_visible_local_constructor_receiver_owner(
-            callable,
-            node,
-            &receiver_name,
-            source,
-            imported_type_bindings,
-            namespace_import_bindings,
-        ) else {
-            return;
-        };
-        let method_col = member_call_method_col(node, source, &method_name);
-        local_receiver_callsites.insert(ReceiverCallSiteKey {
-            receiver_name: receiver_name.clone(),
-            method_name: method_name.clone(),
-            line: Some(node.start_position().row as u32 + 1),
-            method_col,
-        });
-        if let Some((owner_name, owner_module)) = owner {
-            edges.push(ManualReceiverCallSpec {
-                source_name: call_source.name.to_string(),
-                source_span: call_source.span,
-                receiver_name,
-                owner_name,
-                owner_module,
-                method_name,
-                method_col,
-                line: Some(node.start_position().row as u32 + 1),
-                allow_global_fallback: false,
-            });
-        }
-    });
-}
-
-fn typescript_visible_local_constructor_receiver_owner(
-    callable: TsNode<'_>,
-    call_node: TsNode<'_>,
-    receiver_name: &str,
-    source: &str,
-    imported_type_bindings: &HashMap<String, ImportedTypeBinding>,
-    namespace_import_bindings: &HashMap<String, String>,
-) -> Option<OptionalReceiverOwnerBinding> {
-    let mut visible_bindings = Vec::new();
-    walk_tree_nodes(callable, &mut |node| {
-        if node.kind() != "variable_declarator"
-            || !receiver_call_belongs_to_callable(node, callable)
-            || node.end_byte() > call_node.start_byte()
-            || !js_ts_local_binding_visible_at_call(node, call_node)
-        {
-            return;
-        }
-        let Some(binding_name) = node
-            .child_by_field_name("name")
-            .and_then(|name_node| trimmed_node_text(name_node, source))
-            .as_deref()
-            .and_then(normalize_parameter_name)
-        else {
-            return;
-        };
-        if binding_name != receiver_name {
-            return;
-        }
-        visible_bindings.push((
-            node.end_byte(),
-            typescript_constructor_receiver_owner(
-                node,
-                callable,
-                source,
-                imported_type_bindings,
-                namespace_import_bindings,
-            ),
-        ));
-    });
-    visible_bindings.sort_by_key(|(end_byte, _)| *end_byte);
-    visible_bindings.pop().map(|(_, owner)| owner)
-}
-
-fn typescript_constructor_receiver_owner(
-    node: TsNode<'_>,
-    callable: TsNode<'_>,
-    source: &str,
-    imported_type_bindings: &HashMap<String, ImportedTypeBinding>,
-    namespace_import_bindings: &HashMap<String, String>,
-) -> OptionalReceiverOwnerBinding {
-    let constructor_type = node
-        .child_by_field_name("value")
-        .and_then(|value_node| typescript_new_expression_constructor_type(value_node, source))?;
-    let owner_name = normalize_type_surface(&constructor_type)?;
-    if typescript_type_import_qualifier(&constructor_type).is_none()
-        && js_ts_visible_local_type_name(callable, node, &owner_name, source)
-    {
-        return Some((owner_name, None));
-    }
-    typescript_receiver_owner_from_type(
-        &constructor_type,
-        imported_type_bindings,
-        namespace_import_bindings,
-    )
-}
-
-fn typescript_new_expression_constructor_type(node: TsNode<'_>, source: &str) -> Option<String> {
-    if node.kind() != "new_expression" {
-        return None;
-    }
-    node.child_by_field_name("constructor")
-        .and_then(|constructor| trimmed_node_text(constructor, source))
-        .map(|constructor| constructor.trim().to_string())
-        .filter(|constructor| !constructor.is_empty())
-}
-
-fn collect_typescript_class_property_receiver_bindings(
-    callable: TsNode<'_>,
-    source: &str,
-    imported_type_bindings: &HashMap<String, ImportedTypeBinding>,
-    namespace_import_bindings: &HashMap<String, String>,
-) -> HashMap<String, ReceiverOwnerBinding> {
-    let mut receiver_bindings = HashMap::new();
-    if callable.kind() != "method_definition" {
-        return receiver_bindings;
-    }
-    let Some(class_node) = enclosing_node_with_kind(callable, &["class_declaration"]) else {
-        return receiver_bindings;
-    };
-    let mut candidates: HashMap<String, Vec<ReceiverOwnerBinding>> = HashMap::new();
-    walk_tree_nodes(class_node, &mut |node| {
-        if node.kind() != "public_field_definition"
-            || !typescript_property_belongs_to_owner(node, class_node)
-        {
-            return;
-        }
-        let Some((field_name, raw_type)) = typescript_property_receiver_binding(node, source)
-        else {
-            return;
-        };
-        let Some(owner) = typescript_receiver_owner_from_type(
-            &raw_type,
-            imported_type_bindings,
-            namespace_import_bindings,
-        ) else {
-            return;
-        };
-        candidates
-            .entry(format!("this.{field_name}"))
-            .or_default()
-            .push(owner);
-    });
-    for (receiver_name, mut owners) in candidates {
-        owners.sort();
-        owners.dedup();
-        if owners.len() == 1 {
-            receiver_bindings.insert(receiver_name, owners.remove(0));
-        }
-    }
-    receiver_bindings
-}
-
-fn typescript_property_belongs_to_owner(property: TsNode<'_>, class_node: TsNode<'_>) -> bool {
-    let mut current = property.parent();
-    while let Some(candidate) = current {
-        if same_ts_span(candidate, class_node) {
-            return true;
-        }
-        if matches!(candidate.kind(), "method_definition" | "class_declaration") {
-            return false;
-        }
-        current = candidate.parent();
-    }
-    false
-}
-
-fn typescript_property_receiver_binding(
-    node: TsNode<'_>,
-    source: &str,
-) -> Option<(String, String)> {
-    let field_name = node
-        .child_by_field_name("name")
-        .and_then(|name| trimmed_node_text(name, source))
-        .as_deref()
-        .and_then(normalize_parameter_name)?;
-    let surface = trimmed_node_text(node, source)?;
-    let head = surface
-        .split('=')
-        .next()
-        .unwrap_or(surface.as_str())
-        .trim_end_matches(';')
-        .trim();
-    let (_, type_side) = head.split_once(':')?;
-    Some((field_name, parameter_type_after_colon(type_side)))
-}
-
-fn typescript_receiver_owner_from_type(
-    raw_type: &str,
-    imported_type_bindings: &HashMap<String, ImportedTypeBinding>,
-    namespace_import_bindings: &HashMap<String, String>,
-) -> OptionalReceiverOwnerBinding {
-    let owner_name = normalize_type_surface(raw_type)?;
-    if let Some(qualifier) = typescript_type_import_qualifier(raw_type) {
-        let module_name = namespace_import_bindings.get(&qualifier)?;
-        return Some((owner_name, Some(module_name.clone())));
-    }
-    if let Some(binding) = imported_type_bindings.get(&owner_name) {
-        return Some((
-            binding.owner_name.clone(),
-            Some(binding.module_name.clone()),
-        ));
-    }
-    Some((owner_name, None))
-}
-
-fn collect_typescript_imported_type_bindings(
-    root: TsNode<'_>,
-    source: &str,
-) -> HashMap<String, ImportedTypeBinding> {
-    let top_level_bindings = collect_typescript_top_level_type_binding_names(root, source);
-    let mut bindings = HashMap::new();
-    let mut duplicates = HashSet::new();
-    let mut cursor = root.walk();
-    for statement in root.named_children(&mut cursor) {
-        if statement.kind() != "import_statement" {
-            continue;
-        }
-        let Some(module_name) = statement
-            .child_by_field_name("source")
-            .and_then(|module| trimmed_node_text(module, source))
-        else {
-            continue;
-        };
-        for (owner_name, local_name) in typescript_import_binding_names(statement, source) {
-            if top_level_bindings.contains(&local_name) {
-                continue;
-            }
-            if duplicates.contains(&local_name) {
-                continue;
-            }
-            if bindings.contains_key(&local_name) {
-                bindings.remove(&local_name);
-                duplicates.insert(local_name);
-                continue;
-            }
-            bindings.insert(
-                local_name,
-                ImportedTypeBinding {
-                    module_name: module_name.clone(),
-                    owner_name,
-                },
-            );
-        }
-    }
-    bindings
-}
-
-fn typescript_import_binding_names(statement: TsNode<'_>, source: &str) -> Vec<(String, String)> {
-    let mut bindings = Vec::new();
-    let mut cursor = statement.walk();
-    for child in statement.named_children(&mut cursor) {
-        if child.kind() != "import_clause" {
-            continue;
-        }
-        let mut import_clause_cursor = child.walk();
-        for clause_child in child.named_children(&mut import_clause_cursor) {
-            match clause_child.kind() {
-                "identifier" => {
-                    if let Some(local_name) = trimmed_node_text(clause_child, source)
-                        .and_then(|name| normalize_parameter_name(&name))
-                    {
-                        bindings.push((local_name.clone(), local_name));
-                    }
-                }
-                "named_imports" => {
-                    let mut named_cursor = clause_child.walk();
-                    for import_specifier in clause_child.named_children(&mut named_cursor) {
-                        if import_specifier.kind() == "import_specifier"
-                            && let Some(binding) =
-                                typescript_import_specifier_binding_names(import_specifier, source)
-                        {
-                            bindings.push(binding);
-                        }
-                    }
-                }
-                _ => {}
-            }
-        }
-    }
-    bindings
-}
-
-fn collect_typescript_namespace_import_bindings(
-    root: TsNode<'_>,
-    source: &str,
-) -> HashMap<String, String> {
-    let top_level_bindings = collect_typescript_top_level_type_binding_names(root, source);
-    let import_local_bindings = collect_typescript_import_local_binding_names(root, source);
-    let mut bindings = HashMap::new();
-    let mut duplicates = HashSet::new();
-    let mut cursor = root.walk();
-    for statement in root.named_children(&mut cursor) {
-        if statement.kind() != "import_statement" {
-            continue;
-        }
-        let Some(module_name) = statement
-            .child_by_field_name("source")
-            .and_then(|module| trimmed_node_text(module, source))
-        else {
-            continue;
-        };
-        for local_name in typescript_namespace_import_names(statement, source) {
-            if top_level_bindings.contains(&local_name)
-                || import_local_bindings.contains(&local_name)
-            {
-                continue;
-            }
-            if duplicates.contains(&local_name) {
-                continue;
-            }
-            if bindings.contains_key(&local_name) {
-                bindings.remove(&local_name);
-                duplicates.insert(local_name);
-                continue;
-            }
-            bindings.insert(local_name, module_name.clone());
-        }
-    }
-    bindings
-}
-
-fn collect_typescript_import_local_binding_names(
-    root: TsNode<'_>,
-    source: &str,
-) -> HashSet<String> {
-    let mut bindings = HashSet::new();
-    let mut cursor = root.walk();
-    for statement in root.named_children(&mut cursor) {
-        if statement.kind() != "import_statement" {
-            continue;
-        }
-        for (_, local_name) in typescript_import_binding_names(statement, source) {
-            bindings.insert(local_name);
-        }
-    }
-    bindings
-}
-
-fn typescript_namespace_import_names(statement: TsNode<'_>, source: &str) -> Vec<String> {
-    let mut names = Vec::new();
-    let mut cursor = statement.walk();
-    for child in statement.named_children(&mut cursor) {
-        if child.kind() != "import_clause" {
-            continue;
-        }
-        let mut import_clause_cursor = child.walk();
-        for clause_child in child.named_children(&mut import_clause_cursor) {
-            if clause_child.kind() != "namespace_import" {
-                continue;
-            }
-            let mut namespace_cursor = clause_child.walk();
-            for namespace_child in clause_child.named_children(&mut namespace_cursor) {
-                if namespace_child.kind() == "identifier"
-                    && let Some(local_name) = trimmed_node_text(namespace_child, source)
-                        .and_then(|name| normalize_parameter_name(&name))
-                {
-                    names.push(local_name);
-                }
-            }
-        }
-    }
-    names
-}
-
-fn typescript_import_specifier_binding_names(
-    import_specifier: TsNode<'_>,
-    source: &str,
-) -> Option<(String, String)> {
-    let name_node = import_specifier.child_by_field_name("name")?;
-    let owner_name =
-        trimmed_node_text(name_node, source).and_then(|name| normalize_parameter_name(&name))?;
-    let local_node = import_specifier
-        .child_by_field_name("alias")
-        .unwrap_or(name_node);
-    let local_name =
-        trimmed_node_text(local_node, source).and_then(|name| normalize_parameter_name(&name))?;
-    Some((owner_name, local_name))
-}
-
-fn collect_typescript_top_level_type_binding_names(
-    root: TsNode<'_>,
-    source: &str,
-) -> HashSet<String> {
-    let mut bindings = HashSet::new();
-    let mut cursor = root.walk();
-    for child in root.named_children(&mut cursor) {
-        collect_typescript_top_level_type_binding_name(child, source, &mut bindings);
-    }
-    bindings
-}
-
-fn collect_typescript_top_level_type_binding_name(
-    node: TsNode<'_>,
-    source: &str,
-    bindings: &mut HashSet<String>,
-) {
-    match node.kind() {
-        "class_declaration"
-        | "interface_declaration"
-        | "type_alias_declaration"
-        | "enum_declaration" => {
-            if let Some(name) =
-                declaration_name(node, source).and_then(|name| normalize_parameter_name(&name))
-            {
-                bindings.insert(name);
-            }
-        }
-        "export_statement" => {
-            let mut cursor = node.walk();
-            for child in node.named_children(&mut cursor) {
-                collect_typescript_top_level_type_binding_name(child, source, bindings);
-            }
-        }
-        _ => {}
-    }
-}
-
-fn collect_typescript_parameter_type_modules(
-    callable: TsNode<'_>,
-    source: &str,
-    namespace_import_bindings: &HashMap<String, String>,
-) -> HashMap<String, String> {
-    let mut receiver_modules = HashMap::new();
-    let Some(parameters) = signature_parameter_surface(callable, source) else {
-        return receiver_modules;
-    };
-    for parameter in split_top_level_parameters(&parameters) {
-        let Some((name_side, type_side)) = parameter.split_once(':') else {
-            continue;
-        };
-        let Some(receiver_name) = parameter_name_before_colon(name_side) else {
-            continue;
-        };
-        let Some(qualifier) =
-            typescript_type_import_qualifier(&parameter_type_after_colon(type_side))
-        else {
-            continue;
-        };
-        let Some(module_name) = namespace_import_bindings.get(&qualifier) else {
-            continue;
-        };
-        receiver_modules.insert(receiver_name, module_name.clone());
-    }
-    receiver_modules
-}
-
-fn typescript_type_import_qualifier(raw_type: &str) -> Option<String> {
-    if raw_type.contains('|') || raw_type.contains('&') {
-        return None;
-    }
-    let mut surface = raw_type.trim().trim_end_matches('?').trim();
-    while let Some(stripped) = surface.strip_prefix("readonly") {
-        surface = stripped.trim_start();
-    }
-    let base = surface
-        .split(['<', '[', '('])
-        .next()
-        .unwrap_or(surface)
-        .trim();
-    let (qualifier, _) = base.rsplit_once('.')?;
-    normalize_parameter_name(qualifier)
 }
 
 fn collect_python_top_level_binding_names(root: TsNode<'_>, source: &str) -> HashSet<String> {
@@ -14467,22 +13885,6 @@ fn ruby_receiver_call(node: TsNode<'_>, source: &str) -> Option<(String, String)
     Some((normalized_receiver_variable(receiver, source)?, method_name))
 }
 
-fn typescript_member_call(node: TsNode<'_>, source: &str) -> Option<(String, String)> {
-    if node.kind() != "call_expression" {
-        return None;
-    }
-    let function = node.child_by_field_name("function")?;
-    if function.kind() != "member_expression" {
-        return None;
-    }
-    let receiver = function.child_by_field_name("object")?;
-    let method = function.child_by_field_name("property")?;
-    Some((
-        normalize_js_ts_private_receiver_surface(&normalized_receiver_variable(receiver, source)?),
-        trimmed_node_text(method, source)?,
-    ))
-}
-
 fn javascript_member_call(node: TsNode<'_>, source: &str) -> Option<(String, String)> {
     if node.kind() != "call_expression" {
         return None;
@@ -17072,16 +16474,7 @@ fn route_language_uses_c_style_comments(language_name: &str) -> bool {
     }
     matches!(
         language_name,
-        "javascript"
-            | "typescript"
-            | "java"
-            | "rust"
-            | "go"
-            | "php"
-            | "csharp"
-            | "dart"
-            | "vue"
-            | "astro"
+        "javascript" | "java" | "rust" | "go" | "php" | "csharp" | "dart" | "vue" | "astro"
     )
 }
 
@@ -20230,7 +19623,6 @@ pub fn index_file(
                                         "csharp_member" => Some(CSHARP_MEMBER_CALLSITE_MARKER),
                                         "php_member" => Some(PHP_MEMBER_CALLSITE_MARKER),
                                         "js_member" => Some(JS_MEMBER_CALLSITE_MARKER),
-                                        "ts_member" => Some(TS_MEMBER_CALLSITE_MARKER),
                                         "dart_member" => Some(DART_MEMBER_CALLSITE_MARKER),
                                         "swift_member" => Some(SWIFT_MEMBER_CALLSITE_MARKER),
                                         _ => callsite_marker,
@@ -25728,13 +25120,22 @@ class Test {
         assert_eq!(rust.graph_query, RUST_GRAPH_QUERY);
         assert_eq!(rust.tags_query, Some(RUST_TAGS_QUERY));
 
+        // TypeScript's rule files moved into `languages::typescript`; the
+        // config must still come back through the same extension lookup, and
+        // TSX must still reuse the tags query it always shared.
         let ts = get_language_for_ext("ts").expect("ts config");
-        assert_eq!(ts.graph_query, TYPESCRIPT_GRAPH_QUERY);
-        assert_eq!(ts.tags_query, Some(TYPESCRIPT_TAGS_QUERY));
+        let ts_row = languages::extraction_for_ext("ts").expect("typescript registry row");
+        assert_eq!(ts.language_name, "typescript");
+        assert_eq!(ts.graph_query, ts_row.graph_query);
+        assert_eq!(ts.tags_query, ts_row.tags_query);
+        assert!(ts.graph_query.contains("ts_member"));
+        assert!(ts.tags_query.is_some());
 
         let tsx = get_language_for_ext("tsx").expect("tsx config");
         assert_eq!(tsx.graph_query, TSX_GRAPH_QUERY);
         assert_eq!(tsx.tags_query, Some(TSX_TAGS_QUERY));
+        assert_eq!(tsx.tags_query, ts.tags_query);
+        assert_ne!(tsx.graph_query, ts.graph_query);
 
         // Kotlin's rule file moved into `languages::kotlin`; the config must
         // still come back through the same extension lookup.
